CU-3332q2  added endpoints in items.py file for items as well as adapted other endpoints
