+++ app/main/controllers/__init__.py
@@ -4,6 +4,7 @@ from flask import jsonify, send_from_directory
 
 from .projects import set_project_routes
 from .sources import set_source_routes
+from .items import set_item_routes
 from .connections import set_connection_routes
 
 
@@ -26,4 +27,5 @@ def set_routes(app):
 
     set_project_routes(app)
     set_source_routes(app)
+    set_item_routes(app)
     set_connection_routes(app)
